Неудачная попытка добавления сервисов.

+++ src/libraries/Libraries.Data/Models/User.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using ThursdayMeetingBot.Libraries.Data.Models.Base;
+using ThursdayMeetingBot.Libraries.Core.Models.Entities;
+using ThursdayMeetingBot.Libraries.Core.Models.Entities.Base;
 
 namespace ThursdayMeetingBot.Libraries.Data.Models
 {
@@ -7,29 +8,29 @@ namespace ThursdayMeetingBot.Libraries.Data.Models
     ///     User.
     /// </summary>
     [Index(nameof(TelegramChatId))]
-    public record User : AggregatedEntity<long>
+    public record User : AggregatedEntity<int>
     {
         /// <summary>
         ///     Unique identifier of user in Telegram.
         /// </summary>
         public int TelegramId { get; set; }
-        
+
         /// <summary>
         ///     User name that starts with "@".
         /// </summary>
         public string UserName { get; set; }
-        
+
         /// <summary>
         ///     User first name.
         /// </summary>
         /// 
         public string FirstName { get; set; }
-        
+
         /// <summary>
         ///     User last name.
         /// </summary>
         public string LastName { get; set; }
-        
+
         /// <summary>
         ///     Unique identifier of telegram chat for current user.
         /// </summary>
